+++ GRF.Tests/GRFTests.cs
@@ -19,24 +19,12 @@ namespace GRF.Tests
             LoadingMode.Immediate,
             LoadingMode.Deferred
         };
-
-        [Test]
-        public void EntryNames_ReturnsEmptyList_BeforeLoadingAFile()
-        {
-            var grf = new Grf();
-            var expected = new List<string>();
-
-            var actual = grf.EntryNames;
-
-            Assert.AreEqual( expected, actual );
-        }
-
+        
         [Test]
         public void EntryNames_ReturnsAllFilesFromTestGrf_AfterLoadingAFile(
             [ValueSource( "InputFiles" )] string inputFile,
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expected = new List<string>() {
                 "data\\0_Tex1.bmp",
                 "data\\11001.txt",
@@ -47,7 +35,7 @@ namespace GRF.Tests
                 "data\\monstertalktable.xml",
                 "data\\resnametable.txt",
                 "data\\t2_¹è°æ1-1.bmp" };
-            grf.Load( inputFile, mode );
+            var grf = Grf.FromFile( inputFile, mode );
 
             var actual = grf.EntryNames;
 
@@ -59,7 +47,6 @@ namespace GRF.Tests
             [ValueSource( "InputFiles" )] string inputFile,
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expectedPaths = new List<string>() {
                 "data\\0_Tex1.bmp",
                 "data\\11001.txt",
@@ -70,7 +57,7 @@ namespace GRF.Tests
                 "data\\monstertalktable.xml",
                 "data\\resnametable.txt",
                 "data\\t2_¹è°æ1-1.bmp" };
-            grf.Load( inputFile, mode );
+            var grf = Grf.FromFile( inputFile, mode );
 
             foreach( var path in expectedPaths )
             {
@@ -85,7 +72,6 @@ namespace GRF.Tests
             [ValueSource( "InputFiles" )] string inputFile,
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expectedPathAndName = new List<(string, string)>() {
                 ("data\\0_Tex1.bmp", "0_Tex1.bmp"),
                 ("data\\11001.txt", "11001.txt"),
@@ -96,7 +82,7 @@ namespace GRF.Tests
                 ("data\\monstertalktable.xml", "monstertalktable.xml"),
                 ("data\\resnametable.txt", "resnametable.txt"),
                 ("data\\t2_¹è°æ1-1.bmp", "t2_¹è°æ1-1.bmp") };
-            grf.Load( inputFile, mode );
+            var grf = Grf.FromFile( inputFile, mode );
 
             foreach( var (path, name) in expectedPathAndName )
             {
@@ -111,7 +97,6 @@ namespace GRF.Tests
             [ValueSource( "InputFiles" )] string inputFile,
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expectedPathAndTypes = new List<(string, string)>() {
                 ("data\\0_Tex1.bmp", "bmp"),
                 ("data\\11001.txt", "txt"),
@@ -122,7 +107,7 @@ namespace GRF.Tests
                 ("data\\monstertalktable.xml", "xml"),
                 ("data\\resnametable.txt", "txt"),
                 ("data\\t2_¹è°æ1-1.bmp", "bmp") };
-            grf.Load( inputFile, mode );
+            var grf = Grf.FromFile( inputFile, mode );
 
             foreach( var (path, type) in expectedPathAndTypes )
             {
@@ -137,8 +122,7 @@ namespace GRF.Tests
             [ValueSource( "InputFiles" )] string inputFile,
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
-            grf.Load( inputFile, mode );
+            var grf = Grf.FromFile( inputFile, mode );
 
             Assert.IsTrue( grf.Count != 0 );
             foreach(var path in grf.EntryNames )
@@ -154,8 +138,7 @@ namespace GRF.Tests
             [ValueSource( "InputFiles" )] string inputFile,
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
-            grf.Load( inputFile, mode );
+            var grf = Grf.FromFile( inputFile, mode );
 
             Assert.IsTrue( grf.Count != 0 );
             foreach( var path in grf.EntryNames )
@@ -168,39 +151,12 @@ namespace GRF.Tests
         }
 
         [Test]
-        public void EntryNames_ReturnsEmptyList_AfterUnloadingAPreviouslyLoadedFile(
-            [ValueSource( "InputFiles" )] string inputFile,
-            [ValueSource( "LoadingModes" )] LoadingMode mode )
-        {
-            var grf = new Grf();
-            var expected = new List<string>();
-            grf.Load( inputFile, mode );
-            grf.Unload();
-
-            var actual = grf.EntryNames;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        public void EntryCount_ReturnsZero_BeforeLoadingAFile()
-        {
-            var grf = new Grf();
-            var expected = 0;
-
-            var actual = grf.Count;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
         public void EntryCount_ReturnsNine_AfterLoadingAFile(
             [ValueSource( "InputFiles" )] string inputFile,
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expected = 9;
-            grf.Load( inputFile, mode );
+            var grf = Grf.FromFile( inputFile, mode );
 
             var actual = grf.Count;
 
@@ -208,79 +164,12 @@ namespace GRF.Tests
         }
 
         [Test]
-        public void EntryCount_ReturnsZero_AfterUnloadingAPreviouslyLoadedFile(
-            [ValueSource( "InputFiles" )] string inputFile,
+        public void Load_ThrowsDirectoryNotFound_WhenPassingInvalidPath(
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
-            var expected = 0;
-            grf.Load( inputFile, mode );
-            grf.Unload();
-
-            var actual = grf.Count;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        public void IsLoaded_ReturnsFalse_BeforeLoadingAFile()
-        {
-            var grf = new Grf();
-            var expected = false;
-
-            var actual = grf.IsLoaded;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        public void IsLoaded_ReturnsTrue_AfterLoadingAFile(
-            [ValueSource( "InputFiles" )] string inputFile,
-            [ValueSource( "LoadingModes" )] LoadingMode mode )
-        {
-            var grf = new Grf();
-            var expected = true;
-            grf.Load( inputFile, mode );
-
-            var actual = grf.IsLoaded;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        public void IsLoaded_ReturnsFalse_AfterUnloadingAPreviouslyLoadedFile(
-            [ValueSource( "InputFiles" )] string inputFile,
-            [ValueSource( "LoadingModes" )] LoadingMode mode )
-        {
-            var grf = new Grf();
-            var expected = false;
-            grf.Load( inputFile, mode );
-            grf.Unload();
-
-            var actual = grf.IsLoaded;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        public void Load_ThrowsDirectoryNotFound_WhenPassingInvalidPath()
-        {
-            var grf = new Grf();
-
-            void throwingMethod() { grf.Load( "some/path/file.grf" ); }
+            void throwingMethod() { Grf.FromFile( "some/path/file.grf", mode ); }
 
             Assert.Throws<DirectoryNotFoundException>( throwingMethod );
-        }
-
-        [Test]
-        public void Signature_ReturnsEmptyString_BeforeLoadingAFile()
-        {
-            var grf = new Grf();
-            var expected = string.Empty;
-
-            var actual = grf.Signature;
-
-            Assert.AreEqual( expected, actual );
         }
 
         [Test]
@@ -288,24 +177,8 @@ namespace GRF.Tests
             [ValueSource( "InputFiles" )] string inputFile,
             [ValueSource( "LoadingModes" )] LoadingMode mode )
         {
-            var grf = new Grf();
             var expected = "Master of Magic";
-            grf.Load( inputFile, mode );
-
-            var actual = grf.Signature;
-
-            Assert.AreEqual( expected, actual );
-        }
-
-        [Test]
-        public void Signature_ReturnsEmptyString_AfterUnloadingAPreviouslyLoadedFile(
-            [ValueSource( "InputFiles" )] string inputFile,
-            [ValueSource( "LoadingModes" )] LoadingMode mode )
-        {
-            var grf = new Grf();
-            var expected = string.Empty;
-            grf.Load( inputFile, mode );
-            grf.Unload();
+            var grf = Grf.FromFile( inputFile, mode );
 
             var actual = grf.Signature;